requirements.txt
numpy
pyppeteer
python-dotenv
opencv-python
pytesseract
requests